+++ app/models/fae/option.rb
@@ -1,6 +1,7 @@
 module Fae
   class Option < ActiveRecord::Base
 
+    include Fae::BaseModelConcern
     include Fae::OptionConcern
 
     validates_inclusion_of :singleton_guard, :in => [0]