+++ app/models/fae/role.rb
@@ -1,5 +1,7 @@
 module Fae
   class Role < ActiveRecord::Base
+
+    include Fae::BaseModelConcern
     include Fae::RoleConcern
 
     has_many :users