+++ CHANGELOG.md
@@ -15,6 +15,7 @@
     + [#49965](https://issues.afinedevelopment.com/issues/49965) Narrow width of color bar
     + [#50794](https://issues.afinedevelopment.com/issues/50794) Industrial JavaScript refactor
     + [#50863](https://issues.afinedevelopment.com/issues/50863) Nested subnav links are highlighted when the page is active
+    + [#40989](https://issues.afinedevelopment.com/issues/50863) Add a configurable change tracker
 - bugs
     + [#48212](https://issues.afinedevelopment.com/issues/48212) Adjust spacing on pages with multiple tables and drawers